requirements.txt
pillow
pillow_heif
ffmpeg-python
rawpy
imageio
tkinterdnd2